requirements.txt
streamlit
pdfplumber
python-docx
docxtpl
pypdf
pdf2image
pillow
openai
tqdm
